refactor: Make each move function a seperate file and change the name of the common folder to moves
