+++ src/moves/avoidGoingBackwards.js
@@ -1,0 +1,9 @@
+export function avoidGoingBackwards(gameState, isMoveSafe) {
+  const myHead = gameState.you.body[0];
+  const myNeck = gameState.you.body[1];
+
+  if (myNeck.x < myHead.x) isMoveSafe.left = false;
+  else if (myNeck.x > myHead.x) isMoveSafe.right = false;
+  else if (myNeck.y < myHead.y) isMoveSafe.down = false;
+  else if (myNeck.y > myHead.y) isMoveSafe.up = false;
+}
