+++ src/moves/avoidWalls.js
@@ -1,0 +1,11 @@
+export function avoidWalls(gameState, isMoveSafe) {
+  const { width: boardWidth, height: boardHeight } = gameState.board;
+  const { x, y } = gameState.you.body[0];
+
+  if (x === 0) isMoveSafe.left = false;
+  if (x === boardWidth - 1) isMoveSafe.right = false;
+  if (y === 0) isMoveSafe.down = false;
+  if (y === boardHeight - 1) isMoveSafe.up = false;
+
+  return isMoveSafe;
+}
